+++ leadflow_agent/services/__init__.py
@@ -5,3 +5,5 @@ from .website_auditor import WebsiteAuditor
 from .browser_auditor import BrowserAuditor
 
 __all__ = ["LeadInvestigator", "WebsiteAuditor", "BrowserAuditor"]
+
+from .visual_auditor import VisualAuditor, VisualAuditOutcome